requirements.txt
aws-cdk-lib
constructs>=10.0.0,<11.0.0
boto3
aws-cdk.aws-lambda-python-alpha
cdk-nag
flake8
pytest
black